backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
httpx
python-jose
pydantic-settings
aiosqlite
pytest
pytest-asyncio
websockets
